various edits.
