+++ msresist/figures/figureM5.py
@@ -15,21 +15,22 @@ from statsmodels.stats.multitest import multipletests
 from .common import subplotLabel, getSetup
 from ..logistic_regression import plotClusterCoefficients, plotROC
 from ..figures.figure2 import plotPCA, plotMotifs, plotDistanceToUpstreamKinase
-from ..pre_processing import MeanCenter
+from ..pre_processing import MeanCenter, filter_NaNpeptides
 
 
 def makeFigure():
     """Get a list of the axis objects and create a figure"""
     # Get list of axis objects
-    ax, f = getSetup((14, 9), (3, 4), multz={0: 1, 4: 1})
+    ax, f = getSetup((15, 12), (4, 4), multz={0: 1, 4: 1, 12:1, 14:1})
 
     # Add subplot labels
     subplotLabel(ax)
 
     # Set plotting format
-    sns.set(style="whitegrid", font_scale=1.2, color_codes=True, palette="colorblind", rc={"grid.linestyle": "dotted", "axes.linewidth": 0.6})
+    sns.set(style="whitegrid", font_scale=1, color_codes=True, palette="colorblind", rc={"grid.linestyle": "dotted", "axes.linewidth": 0.6})
 
     X = pd.read_csv("msresist/data/MS/CPTAC/CPTAC-preprocessedMotfis.csv").iloc[:, 1:]
+    X = filter_NaNpeptides(X, tmt=2)
 
     with open('msresist/data/pickled_models/binomial/CPTACmodel_BINOMIAL_CL24_W15_TMT2', 'rb') as p:
         model = pickle.load(p)[0]
@@ -65,16 +66,64 @@ def makeFigure():
 
     # plot Upstream Kinases
     plotDistanceToUpstreamKinase(model, [11, 12], ax[6], num_hits=3)
+    plot_NetPhoresScoreByKinGroup("msresist/data/cluster_analysis/CPTAC_NK_C12.csv", ax[7], n=5, title="Cluster 12 Kinase Predictions")
 
-    # Cluster 11
-    plot_GO(11, ax[7], n=5, title="GO Cluster 11")
+    # GO Cluster 11
+    # plot_GO(11, ax[8], n=5, title="GO Cluster 11")
 
-    # Cluster 12
-    plot_GO(12, ax[8], n=3, title="GO Cluster 12")
+    # GO Cluster 12
+    # plot_GO(12, ax[9], n=3, title="GO Cluster 12")
 
-    plot_NetPhoresScoreByKinGroup("msresist/data/cluster_analysis/CPTAC_NK_C12.csv", ax[9], n=5, title="Cluster 12 Kinase Predictions")
+    # Peptides Cluster 11
+    y = pd.DataFrame(centers["Type"]).reset_index()
+    y.columns = ["Sample.ID", "Type"]
+    X["cluster"] = model.labels()
+    c11 = X[X["cluster"] == 11].drop("cluster", axis=1)
+    d = {"PEAK1": "Y635-p", "ARHGEF7": "S703-p", "PAK4": "S181-p", "FLNA": "S2128-p", "PTPN11": "Y546-p", "HBA2": "T68-p", "HBB": "T88-p", "HBD":"S73-p", "HBG1": "S140-p"}
+    plotPeptidesByFeature(c11, y, d, ["Type", "Tumor", "NAT"], ax[8], title="Gas Transport & Cytoskletal remodeling")
+
+    # Peptides Cluster 12
+    c12 = X[X["cluster"] == 12].drop("cluster", axis=1)
+    d = {"MCM4": "S105-p", "MCM3": "T722-p", "TP53BP1": "T1672-p", "MCM4": "S105-p", "BRCA1": "S114-p", "ATRX":"S1348-p", "CDK1": "Y15-p;T14-p", "CDK12": "S102-p;S105-p", "CDK13": "S317-p", "CDK16": "S119-p", "CENPF": "T2997-p"}
+    plotPeptidesByFeature(c12, y, d, ["Type", "Tumor", "NAT"], ax[9], title="DNA Damage")
+
+    # String plots
+    ax[10].axis("off")
+    ax[11].axis("off")
 
     return f
+
+
+def plotPeptidesByFeature(X, y, d, feat_labels, ax, loc='best', title=False):
+    """Plot and compare specific peptides by feature."""
+    x = X.set_index(["Gene", "Position"])
+    n = list(d.keys())
+    p = list(d.values())
+    dfs = []
+    for i in range(len(n)):
+        dfs.append(pd.DataFrame(x.loc[n[i], p[i]]).T)
+    c = pd.concat(dfs).reset_index()
+    c.columns = ["Gene", "Position"] + list(c.columns[2:])
+
+    # Farmat data to concatenate feature
+    c["SeqPos"] = [s + ";" + c["Position"].iloc[i] for i, s in enumerate(c["Gene"])]
+    c = c.set_index("SeqPos").T.iloc[4:, :].reset_index()
+
+    assert np.all(list(c["index"]) == list(y["Sample.ID"]))
+    c = c.reset_index().iloc[:, 2:]
+
+    # Add feature
+    f1, f2, f3 = feat_labels
+    c[f1] = y.iloc[:, 1].values
+    c[f1] = c[f1].replace(0, f2)
+    c[f1] = c[f1].replace(1, f3)
+
+    dm = pd.melt(c, id_vars=f1, value_vars=c.columns[:-1], var_name="p-site", value_name="mean log(p-signal)")
+
+    sns.barplot(data=dm, x=f1, y="mean log(p-signal)", hue="p-site", ci=None, ax=ax)
+    ax.legend(prop={"size": 8}, loc=loc)
+    if title:
+        ax.set_title(title)
 
 
 def plot_NetPhoresScoreByKinGroup(PathToFile, ax, n=5, title=False):
