+++ msresist/figures/figureM6.py
@@ -10,7 +10,7 @@ from sklearn.preprocessing import StandardScaler
 from ..pre_processing import filter_NaNpeptides
 from .figure2 import plotDistanceToUpstreamKinase
 from .figureM4 import find_patients_with_NATandTumor
-from .figureM5 import plot_clusters_binaryfeatures, build_pval_matrix, calculate_mannW_pvals, plot_GO
+from .figureM5 import plot_clusters_binaryfeatures, build_pval_matrix, calculate_mannW_pvals, plot_GO, plotPeptidesByFeature
 from ..logistic_regression import plotROC, plotClusterCoefficients
 from .common import subplotLabel, getSetup
 
@@ -65,7 +65,7 @@ def makeFigure():
 
     # Logistic Regression
     centers["STK11"] = y_["STK11.mutation.status"].values
-    lr = LogisticRegressionCV(Cs=10, cv=10, solver="saga", max_iter=100000, tol=1e-4, n_jobs=-1, penalty="l1", class_weight="balanced")
+    lr = LogisticRegressionCV(cv=5, solver="saga", max_iter=100000, tol=1e-4, n_jobs=-1, penalty="elasticnet", l1_ratios=[0.1])
     plotROC(ax[1], lr, centers.iloc[:, :-1].values, centers["STK11"], cv_folds=4, title="ROC STK11")
     ax[1].legend(loc='lower right', prop={'size': 8})
     plotClusterCoefficients(ax[2], lr.fit(centers.iloc[:, :-1], centers["STK11"].values), list(centers.columns[:-1]), title="STK11")
@@ -88,36 +88,3 @@ def makeFigure():
     plotPeptidesByFeature(c8, y, d, ["STK11 status", "WT", "Mutant"], ax=ax[6], title="Golgi Fragmentation peptides")
 
     return f
-
-
-def plotPeptidesByFeature(X, y, d, feat_labels, ax, loc='best', title=False):
-    """Plot and compare specific peptides by feature. Input data should contain a column with the
-    feature of interest"""
-    x = X.set_index(["Gene", "Position"])
-    n = list(d.keys())
-    p = list(d.values())
-    dfs = []
-    for i in range(len(n)):
-        dfs.append(pd.DataFrame(x.loc[n[i], p[i]]).T)
-    c = pd.concat(dfs).reset_index()
-    c.columns = ["Gene", "Position"] + list(c.columns[2:])
-
-    # Farmat data to concatenate feature
-    c["SeqPos"] = [s + ";" + c["Position"].iloc[i] for i, s in enumerate(c["Gene"])]
-    c = c.set_index("SeqPos").T.iloc[4:, :].reset_index()
-
-    assert np.all(list(c["index"]) == list(y["Sample.ID"]))
-    c = c.reset_index().iloc[:, 2:]
-
-    # Add feature
-    f1, f2, f3 = feat_labels
-    c[f1] = y.iloc[:, 1].values
-    c[f1] = c[f1].replace(0, f2)
-    c[f1] = c[f1].replace(1, f3)
-
-    dm = pd.melt(c, id_vars=f1, value_vars=c.columns[:-1], var_name="p-site", value_name="mean log(p-signal)")
-
-    sns.barplot(data=dm, x=f1, y="mean log(p-signal)", hue="p-site", ci=None, ax=ax)
-    ax.legend(prop={"size": 8}, loc=loc)
-    if title:
-        ax.set_title(title)
